Add Dockerfile for Antora setup
Created a new Dockerfile to set up an Antora environment for documentation generation. This file uses the latest Antora image as the base and installs essential dependencies such as Git, OpenSSH, Curl, among others. The Dockerfile also defines '/documents' as the working directory and sets Antora as the entry point. The setup enhances reproducibility and consistency in documentation generation across different environments.